refactor(scheduler): hoist the extension contract into pkg/scheduling
The extension contract (SchedulerExtension, SnapshotInput,
PostAssignInput, ...) was duplicated in scheduling/v1 and
scheduling/v1alpha, so an extension written against one package's types
could not be registered on the other implementation's pool — downstream
consumers (hatchet-cloud's autoscaling extension) broke against the
scheduling.Pool interface and could never have followed the
per-shard scheduler flag.

Define the contract once in pkg/scheduling, alias it from both
implementations (existing DTO alias pattern; downstream code referencing
v1.SchedulerExtension etc. keeps compiling unchanged), and expose
AddExtension on the Pool interface so extensions register against
whichever implementation the shard runs. The loader now registers the
prometheus extension once, outside the flag branch.
